+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/pyaud/compare/v3.2.8...HEAD)
+[Unreleased](https://github.com/jshwi/pyaud/compare/v3.2.9...HEAD)
+------------------------------------------------------------------------
+
+[3.2.9](https://github.com/jshwi/pyaud/releases/tag/v3.2.9) - 2021-11-08
 ------------------------------------------------------------------------
 ### Fixed
 Bypasses `TypeError` when configuring logger